Extracted SubHeadingTitleWithConnections
